do not set/reset Better Pawn Control policy if none is configured
There are other ways to set BPC policies, such as BPC's own alert state.
With that this mod's handling becomes superfluous and in fact
an additional chore, as it is necessary to explicitly unset this mod's
policy state, and it must be done before reverting BPC's, otherwise this
mod would set the battle policy again.